feat: cleanup-ui-factory-session-detail-test-split-001 - Preserve loading, success, and durable summary coverage
