+++ ui/src/features/factory-session-detail/components/factory-session-detail-panel.test.tsx
@@ -8,7 +8,11 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { FactoryOrchestratorKind } from "../../../api/generated/openapi";
 import { FactorySessionDetailPanel } from "./factory-session-detail-panel";
 import {
-  createDeferred,
+  BASELINE_DISPATCH_ID,
+  BASELINE_SESSION_ID,
+  mockJavaScriptSessionBetaFetchWithDeferredDispatch,
+} from "./test-support/factory-session-detail-panel.baseline-fixtures";
+import {
   jsonResponse,
   renderWithQueryClient,
 } from "./test-support/factory-session-detail-panel.test-helpers";
@@ -22,149 +26,21 @@ describe("FactorySessionDetailPanel", () => {
     vi.unstubAllGlobals();
   });
 
-  it("shows loading and success states for a JavaScript factory session", async () => {
-    const dispatchDetailResponse = createDeferred<Response>();
-    vi.mocked(globalThis.fetch).mockImplementation(async (input) => {
-      const url = String(input);
-      if (url.endsWith("/factory-sessions/session-beta")) {
-        return jsonResponse({
-          factoryDir: "/workspace/root/beta",
-          folderPath: "/workspace/root",
-          id: "session-beta",
-          isDefault: false,
-          project: "beta",
-          runtime: {
-            artifacts: [
-              {
-                id: "artifact-1",
-                kind: "CHILD_RESULT",
-                label: "review output",
-                visibility: "CUSTOMER",
-              },
-            ],
-            dispatches: [
-              {
-                dispatchKind: "JAVASCRIPT_AGENT",
-                id: "dispatch-1",
-                javascript: {
-                  executionMode: " live ",
-                  taskKind: "AGENT",
-                  taskLabel: " Review child task ",
-                },
-                orchestratorKind: FactoryOrchestratorKind.JAVASCRIPT,
-                providerSessionRefs: [
-                  {
-                    id: "provider-session-1",
-                    kind: "session_id",
-                    provider: "codex",
-                  },
-                ],
-                sessionId: "session-beta",
-                status: "COMPLETED",
-                warnings: [
-                  {
-                    code: "DISPATCH_WARNING",
-                    message: "child agent retry scheduled",
-                  },
-                ],
-              },
-            ],
-            javascript: {
-              checkpoints: [
-                {
-                  id: "cp-1",
-                  label: "plan",
-                  summary: "saved plan checkpoint",
-                },
-              ],
-              childDispatchCounts: {
-                completed: 4,
-                queued: 1,
-                running: 2,
-              },
-              phase: "review",
-              phases: ["plan", "review"],
-              scriptStatus: "IDLE",
-            },
-            lifecycle: {
-              startedAt: "2026-06-08T14:00:00Z",
-              updatedAt: "2026-06-08T14:05:00Z",
-            },
-            orchestratorKind: FactoryOrchestratorKind.JAVASCRIPT,
-            progress: {
-              categories: {},
-              factoryState: "RUNNING",
-              inFlightCount: 0,
-              totalTokens: 0,
-            },
-            status: "IDLE",
-            usage: { resources: [] },
-          },
-          target: { kind: "named", name: "beta" },
-        });
-      }
-      if (url.endsWith("/factory-sessions/session-beta/result")) {
-        return jsonResponse({
-          resultArtifactRef: {
-            id: "artifact-final",
-            kind: "FINAL_RESULT",
-            visibility: "CUSTOMER",
-          },
-          sessionId: "session-beta",
-          status: "IDLE",
-        });
-      }
-      if (url.endsWith("/factory-sessions/session-beta/partial-result")) {
-        return jsonResponse({
-          partialResultArtifactRef: {
-            id: "artifact-partial",
-            kind: "CHILD_RESULT",
-            visibility: "CUSTOMER",
-          },
-          phase: "review",
-          sessionId: "session-beta",
-        });
-      }
-      if (
-        url.endsWith("/factory-sessions/session-beta/dispatches/dispatch-1")
-      ) {
-        return dispatchDetailResponse.promise;
-      }
-      return new Response("not found", { status: 404 });
-    });
+  it("shows loading and success states for a JavaScript factory session dispatch drilldown", async () => {
+    const dispatchDetailResponse =
+      mockJavaScriptSessionBetaFetchWithDeferredDispatch();
 
     renderWithQueryClient(
-      <FactorySessionDetailPanel sessionID="session-beta" />,
-    );
-
-    expect(screen.getByRole("status").textContent).toContain(
-      "Loading factory session runtime…",
+      <FactorySessionDetailPanel sessionID={BASELINE_SESSION_ID} />,
     );
 
     await waitFor(() => {
       expect(screen.getByText("Runtime")).toBeTruthy();
     });
 
-    expect(screen.getByText("JavaScript workflow")).toBeTruthy();
-    expect(screen.getAllByText("Idle").length).toBeGreaterThanOrEqual(1);
-    expect(screen.getByText("review")).toBeTruthy();
-    expect(
-      screen.getByText("cp-1 (plan) — saved plan checkpoint"),
-    ).toBeTruthy();
-    expect(screen.getByText("child agent retry scheduled")).toBeTruthy();
-    expect(screen.getByText("artifact-final · FINAL_RESULT")).toBeTruthy();
-    expect(screen.getByText("artifact-partial · CHILD_RESULT")).toBeTruthy();
-    expect(screen.getByText("Execution mode: live")).toBeTruthy();
-    expect(
-      screen.getByText(
-        "Provider session: codex / session_id / provider-session-1",
-      ),
-    ).toBeTruthy();
-    expect(screen.queryByText("rawCheckpointBody")).toBeNull();
-
     const user = userEvent.setup();
     const drilldownTrigger = screen.getByRole("button", {
-      name: "Expand dispatch detail for dispatch-1",
+      name: `Expand dispatch detail for ${BASELINE_DISPATCH_ID}`,
     });
 
     expect(drilldownTrigger.getAttribute("aria-expanded")).toBe("false");
@@ -238,52 +114,6 @@ describe("FactorySessionDetailPanel", () => {
     expect(screen.getByText("$0.21")).toBeTruthy();
     expect(screen.getByText("4,400 ms")).toBeTruthy();
     expect(screen.getByText("Token budget was nearly exhausted.")).toBeTruthy();
-  });
-
-  it("shows loading state while factory session runtime data is pending", async () => {
-    let resolveFetch: (value: Response) => void = () => undefined;
-    const pendingResponse = new Promise<Response>((resolve) => {
-      resolveFetch = resolve;
-    });
-    vi.mocked(globalThis.fetch).mockReturnValue(pendingResponse);
-
-    renderWithQueryClient(
-      <FactorySessionDetailPanel sessionID="session-beta" />,
-    );
-
-    expect(screen.getByText("Loading factory session runtime…")).toBeTruthy();
-    expect(screen.queryByText("Dynamic workflow (JavaScript factory session)")).toBeNull();
-    expect(screen.queryByText("This factory session is no longer available.")).toBeNull();
-
-    resolveFetch(
-      jsonResponse({
-        factoryDir: "/workspace/root/beta",
-        folderPath: "/workspace/root",
-        id: "session-beta",
-        isDefault: false,
-        project: "beta",
-        runtime: {
-          lifecycle: {
-            startedAt: "2026-06-08T14:00:00Z",
-            updatedAt: "2026-06-08T14:05:00Z",
-          },
-          orchestratorKind: FactoryOrchestratorKind.PETRI,
-          progress: {
-            categories: {},
-            factoryState: "RUNNING",
-            inFlightCount: 0,
-            totalTokens: 0,
-          },
-          status: "IDLE",
-          usage: { resources: [] },
-        },
-        target: { kind: "named", name: "beta" },
-      }),
-    );
-
-    await waitFor(() => {
-      expect(screen.queryByText("Loading factory session runtime…")).toBeNull();
-    });
   });
 
   it("shows a not-found state when the factory session is no longer available", async () => {
